Little perfomance improvement
